Fixed some stuff, broke some stuff, life goes on...

+++ Kernel/util/new.cpp
@@ -7,13 +7,11 @@
 
 #include "new.h"
 
-#include <util/alloc.h>
-#include <cstdint>
 
-void* operator new(uint32_t size) {
+void* operator new(size_t size) {
 	return malloc(size);
 }
-void* operator new[](uint32_t size) {
+void* operator new[](size_t size) {
 	return malloc(size);
 }
 void operator delete(void* block) {
