+++ Kernel/util/new.h
@@ -8,10 +8,11 @@
 #ifndef UTIL_NEW_H_
 #define UTIL_NEW_H_
 
+#include <stdlib.h>
 #include <cstdint>
 
-void* operator new(uint32_t size);
-void* operator new[](uint32_t size);
+void* operator new(size_t size);
+void* operator new[](size_t size);
 void operator delete(void* block);
 void operator delete[](void* block);
 #endif /* UTIL_NEW_H_ */
